# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Task list
  description: >-
    This is an OpenAPI Specification created by oas-wizard
    (https://github.com/pafmon/oas-wizard)  
  contact:
    name: PabloFM
    email: pablofm@us.es
    url: 'http://www.isa.us.es'
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
paths:
...
  /api/v1/tags/modify:
    post:
      summary: Modify tags for tasks
      operationId: modifyTags
      x-router-controller: apiv1tagsController
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
                taskIds:
                  type: array
                  items:
                    type: string
                operation:
                  type: string
                  enum: [tag, untag]
      responses:
        '200':
          description: Tags modified successfully
...